[IMP] zen: verses about code complexity
Writing beautiful code is difficult because often the problems are
complicated. When the code gets as complicated as the problems it tries
to solve then there is a chance that the code gets ugly.

Because beautiful is better than ugly, and that complicated code often
looks ugly. One should write simpler code instead.

There are situation where simple solutions are not enough, for those it
is ok to add a bit of complexity, as long as that complexity is well
managed.