[IMP] zen: continue reflexion on ambiguity
Any good python peom should address the need to be concise and clear.

Thus we try to sensibilise the reader on the fact that when one is not sure,
the temptation is to just guess what a certain line of code means or assume
what the author was aiming for.

One should always go to the source, ask either the original author or the
maintainer for more information. Barring that, test the code, test your
assumptions and try to break the code with edge cases.
Better be safe than sorry.