[IMP] zen: add verse about code ambiguity
A verse about how to deal with code ambiguity was missing to the zen poem.

Computers have made humans superstitious: To exorcise the demons in our
computers we perform the sacred ritual of turning them off and then on.
Supposedly this will fix any mysterious problem. However, computers are
not magic. If your code is not working, there is a reason and only
careful, critical thinking will solve it. Refuse the temptation to
blindly try solutions until something seems to work; often you have
merely masked the problem rather than solved it.

task-chakra-wide-open