[IMP] zen: verse about code ambiguity
Computers do only what we tell them to do: if your code is not behaving as you would like it
to, it is because it is doing what you have told it to do. Trying to fix the behavior by blindly
attempting several different solutions until one works is a poor strategy - you may end up
masking the problem rather than solving it. Resist the temptation. Instead, think through the
logic of the problem and apply critical thinking to come up with an appropriate solution.