[IMP] zen: Add new verse about ambiguity
Writing clean code is good, but there should also be some advice for
people who want to read and reuse the code.

Ambiguity is a common factor of confusion, and has the effect of causing
disparity beween the intended use of code and its actual usage.

A small verse about ambiguity should help as a warning.