[IMP] zen: verse about guessing
The verse adresses the temptation of guessing, in a difficult situation

When the code is not working, there must be an error in the logic we implemented.
Hence, only correct logic can fix the error. We should never try to blindly guess
why the code isn’t working. It will only lead to time waste. We should always
debug the code in a proper manner to rectify the error.