[FIX] zen: verses about errors fixing strategy
It's super important to understant the reason behind the error, because
not understanding the root cause of the problem and blindly try random
solutions till one of them work might probably mask the origianl problem
and not solving it.

Afterall, if we don't understand the source of the problem, are we really
improving ourselves?? It's up to you to decide, but be responsible!