[IMP] zen: verses about critical thinking
Having not working code is ok, but this means the code has a problem.
In this ambigous situation, we can't guess different solutions blindly
trying to bypass the current issue. This will be masking the problem
without trying to understand and fix the root issue of it, so instead
we need to think critically about it to find the best solution that
can solve the root issue of the problem.