[IMP] zen: add a verse about temptation to guess
Why the commit :
To add the line about responsability, and ask ourself the right quesiton at the moment of coding.

That line was added at that place since it's a poem from the internet :
original document at [PEP 20].
new entries should follow the
same order as in the original document.

[PEP 20]: https://peps.python.org/pep-0020/