[IMP] zen.txt: complete poem with final line
The poem of the zen.txt file was good, but was missing many lines,
making it incomplete [1]. To improve the poem for a better impact,
a new line from the original poem was added.

Developpers tend to try random fixes that would be likely to create
others issues. This last line will make developpers more likely to
search for a real answer rather than trying until something works.
This will likely reduce the need for fixes in the future.

-----------------------------------------------------------------------

[1]: https://realpython.com/zen-of-python/