[ADD] zen: verses about handling ambiguity
The easiest reaction when facing ambiguity, in life or code alike, is the
tendancy to attempt to guess an answer that lifts you from the state of
confusion and -well, ambiguity-, to a state of clarity and certainty.

A good coding practice is to avoid mere guesses, and look deeper into the
troubling code, get comfortable with it, and explore what it has to tell you
about its actual truth.