[IMP] zen: verse about enhancing clarity
A beautiful code is a code that is clear, one that is not abigious in any sort of way, a code that's easily read and clear to the reader.
A clear code is also a piece of code that is easily read by others.
Clarity is important as it makes it easier to understand.