[IMP] zen: verse about code ambiguity
Writing beautiful code is difficult because often the problems are ambigious.
When the code gets as ambigious as the problems it tries to solve then there
is a chance that the code gets ugly.

Because beautiful is better than ugly, and that ambigious code often looks ugly.

There are situation where simple solutions are not enough, for those it is ok
to add a bit of ambiguity, as long as that complexity is well managed.